+++ Sources/DeckKit/Favorites/FavoriteContext.swift
@@ -12,26 +12,32 @@ import SwiftUI
 ///
 /// The class uses a ``UserDefaultsFavoriteService`` to store state by
 /// default, but you can use any custom service.
-public class FavoriteContext<Item: Identifiable, Service: FavoriteService>: ObservableObject, FavoriteService where Service.Item == Item {
+public class FavoriteContext<Item: Identifiable>: ObservableObject {
 
     /// Create a default context instance.
     ///
     /// This instance uses a ``UserDefaultsFavoriteService``.
-    public init() where Service == UserDefaultsFavoriteService<Item> {
-        self.service = UserDefaultsFavoriteService<Item>()
-        self.favorites = getFavorites()
+    public convenience init() {
+        self.init(service: UserDefaultsFavoriteService<Item>())
     }
 
     /// Create a context with a custom service.
     ///
     /// - Parameters:
     ///   - service: The service to use to manage favorites.
-    public init(service: Service) where Item == Service.Item {
-        self.service = service
-        self.favorites = getFavorites()
+    public init<Service: FavoriteService>(service: Service) where Service.Item == Item {
+        self.favorites = service.getFavorites()
+        self._getFavorites = service.getFavorites
+        self._isFavorite = service.isFavorite
+        self._setIsFavorite = service.setIsFavorite
+        self._toggleIsFavorite = service.toggleIsFavorite
     }
 
-    private let service: Service
+    private let _getFavorites: () -> [Item.ID]
+    private let _isFavorite: (Item) -> Bool
+    private let _setIsFavorite: (Bool, Item) -> Void
+    private let _toggleIsFavorite: (Item) -> Void
+
 
     /// The item IDs that are currently marked as favorites.
     @Published
@@ -51,23 +57,23 @@ public extension FavoriteContext {
 
     /// Get all favorites.
     func getFavorites() -> [Item.ID] {
-        service.getFavorites()
+        _getFavorites()
     }
     
     /// Check if a certain item is marked as a favorite.
     func isFavorite(_ item: Item) -> Bool {
-        service.isFavorite(item)
+        _isFavorite(item)
     }
     
     /// Set the favorite state of a certain item.
     func setIsFavorite(_ isFavorite: Bool, for item: Item) {
-        service.setIsFavorite(isFavorite, for: item)
+        _setIsFavorite(isFavorite, item)
         favorites = getFavorites()
     }
     
     /// Toggle the favorite state of a certain item.
     func toggleIsFavorite(for item: Item) {
-        service.toggleIsFavorite(for: item)
+        _toggleIsFavorite(item)
         favorites = getFavorites()
     }
 }